backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client for AgentForge integration
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0

# AgentForge dependencies (will be added when integrated)
# langdetect
# crewai
# crewai_tools
# pyyaml
# python-dotenv
